+++ Essential 8 Knowledge Base/Essential_8_Knowledge_BaseApp.swift
@@ -12,6 +12,7 @@ struct Essential_8_Knowledge_BaseApp: App {
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .environmentObject(ProgressStore.shared)
         }
     }
 }